+++ src/main/java/weixin/popular/api/VehicleAPI.java
@@ -7,14 +7,14 @@ import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
 import org.apache.http.entity.StringEntity;
 
-import weixin.popular.bean.vehicle.EntranceNotification;
-import weixin.popular.bean.vehicle.EntranceNotificationResult;
+import weixin.popular.bean.vehicle.PartnerpayNotification;
+import weixin.popular.bean.vehicle.PartnerpayNotificationResult;
 import weixin.popular.bean.vehicle.NavigateToAppForAuth;
 import weixin.popular.bean.vehicle.NavigateToAppForAuthExtraData;
-import weixin.popular.bean.vehicle.PayApply;
-import weixin.popular.bean.vehicle.PayApplyResult;
-import weixin.popular.bean.vehicle.QueryState;
-import weixin.popular.bean.vehicle.QueryStateResult;
+import weixin.popular.bean.vehicle.PartnerpayPayApply;
+import weixin.popular.bean.vehicle.PartnerpayPayApplyResult;
+import weixin.popular.bean.vehicle.PartnerpayQueryState;
+import weixin.popular.bean.vehicle.PartnerpayQueryStateResult;
 import weixin.popular.client.LocalHttpClient;
 import weixin.popular.util.JsonUtil;
 import weixin.popular.util.MapUtil;
@@ -22,12 +22,15 @@ import weixin.popular.util.SignatureUtil;
 import weixin.popular.util.XMLConverUtil;
 
 /**
- * 车主服务
+ * 车主服务，只实现车主服务特有的接口，其他支付相关的接口和普通微信支付一样，调用PayMchAPI即可
+ * 
+ * 接口文档
+ * https://pay.weixin.qq.com/wiki/doc/api/vehicle_v2_sl.php?chapter=2_3&index=1
  * 
  * @author around
  *
  */
-public class PayVeticleAPI extends BaseAPI {
+public class VehicleAPI extends BaseAPI {
 
 	/**
 	 * 入场通知
@@ -36,7 +39,7 @@ public class PayVeticleAPI extends BaseAPI {
 	 * @param key
 	 * @return
 	 */
-	public static EntranceNotificationResult entranceNotification(EntranceNotification notification, String key) {
+	public static PartnerpayNotificationResult partnerpayNotification(PartnerpayNotification notification, String key) {
 		Map<String, String> map = MapUtil.objectToMap(notification);
 		if (notification.getScene_info() != null) {
 			map.put("scene_info", "{\"scene_info\":" + JsonUtil.toJSONString(notification.getScene_info()) + "}");
@@ -44,9 +47,11 @@ public class PayVeticleAPI extends BaseAPI {
 		String sign = SignatureUtil.generateSign(map, notification.getSign_type(), key);
 		notification.setSign(sign);
 		String xml = XMLConverUtil.convertToXML(notification);
-		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader).setUri(MCH_URI + "/vehicle/partnerpay/notification")
+		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader)
+				.setUri(MCH_URI + "/vehicle/partnerpay/notification")
 				.setEntity(new StringEntity(xml, Charset.forName("utf-8"))).build();
-		return LocalHttpClient.executeXmlResult(httpUriRequest, EntranceNotificationResult.class, notification.getSign_type(), key);
+		return LocalHttpClient.executeXmlResult(httpUriRequest, PartnerpayNotificationResult.class,
+				notification.getSign_type(), key);
 	}
 
 	/**
@@ -56,7 +61,7 @@ public class PayVeticleAPI extends BaseAPI {
 	 * @param key
 	 * @return
 	 */
-	public static PayApplyResult payApply(PayApply apply, String key) {
+	public static PartnerpayPayApplyResult partnerpayPayApply(PartnerpayPayApply apply, String key) {
 		Map<String, String> map = MapUtil.objectToMap(apply);
 		if (apply.getScene_info() != null) {
 			map.put("scene_info", "{\"scene_info\":" + JsonUtil.toJSONString(apply.getScene_info()) + "}");
@@ -64,30 +69,53 @@ public class PayVeticleAPI extends BaseAPI {
 		String sign = SignatureUtil.generateSign(map, apply.getSign_type(), key);
 		apply.setSign(sign);
 		String xml = XMLConverUtil.convertToXML(apply);
-		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader).setUri(MCH_URI + "/vehicle/partnerpay/payapply").setEntity(new StringEntity(xml, Charset.forName("utf-8")))
-				.build();
-		return LocalHttpClient.executeXmlResult(httpUriRequest, PayApplyResult.class, apply.getSign_type(), key);
+		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader)
+				.setUri(MCH_URI + "/vehicle/partnerpay/payapply")
+				.setEntity(new StringEntity(xml, Charset.forName("utf-8"))).build();
+		return LocalHttpClient.executeXmlResult(httpUriRequest, PartnerpayPayApplyResult.class, apply.getSign_type(),
+				key);
 	}
 
 	/**
-	 * 查询
+	 * 查询用户状态
 	 * 
 	 * @param query
 	 * @param key
 	 * @return
 	 */
-	public static QueryStateResult queryState(QueryState query, String key) {
+	public static PartnerpayQueryStateResult partnerpayQueryState(PartnerpayQueryState query, String key) {
 		Map<String, String> map = MapUtil.objectToMap(query);
 		String sign = SignatureUtil.generateSign(map, query.getSign_type(), key);
 		query.setSign(sign);
 		String xml = XMLConverUtil.convertToXML(query);
-		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader).setUri(MCH_URI + "/vehicle/partnerpay/querystate")
+		HttpUriRequest httpUriRequest = RequestBuilder.post().setHeader(xmlHeader)
+				.setUri(MCH_URI + "/vehicle/partnerpay/querystate")
 				.setEntity(new StringEntity(xml, Charset.forName("utf-8"))).build();
-		return LocalHttpClient.executeXmlResult(httpUriRequest, QueryStateResult.class, query.getSign_type(), key);
+		return LocalHttpClient.executeXmlResult(httpUriRequest, PartnerpayQueryStateResult.class, query.getSign_type(),
+				key);
 	}
 
-	public static NavigateToAppForAuth navigateToMappForAuth(String path, String tradeScene, String appid, String subAppid, String mchid, String subMchid, String openid, String subOpenid,
-			String plateNumber, String materialInfo, String channelType, String key) {
+	/**
+	 * <a href=
+	 * "https://pay.weixin.qq.com/wiki/doc/api/pap_sl_jt_v2.php?chapter=20_932&index=10">跳转到小程序/H5/APP，用户授权/开通接口api</a>
+	 * 
+	 * @param path
+	 * @param tradeScene
+	 * @param appid
+	 * @param subAppid
+	 * @param mchid
+	 * @param subMchid
+	 * @param openid
+	 * @param subOpenid
+	 * @param plateNumber
+	 * @param materialInfo
+	 * @param channelType
+	 * @param key
+	 * @return
+	 */
+	public static NavigateToAppForAuth navigateToAppForAuth(String path, String tradeScene, String appid,
+			String subAppid, String mchid, String subMchid, String openid, String subOpenid, String plateNumber,
+			String materialInfo, String channelType, String key) {
 		NavigateToAppForAuth nav = new NavigateToAppForAuth();
 		nav.setPath(path);
 		NavigateToAppForAuthExtraData extraData = new NavigateToAppForAuthExtraData();
@@ -100,9 +128,7 @@ public class PayVeticleAPI extends BaseAPI {
 		extraData.setSub_openid(subOpenid);
 		extraData.setChannel_type(channelType);
 		extraData.setMaterial_info(materialInfo);
-		if ("HIGHWAY".equalsIgnoreCase(tradeScene)) {
-			extraData.setPlate_number(plateNumber);
-		}
+		extraData.setPlate_number(plateNumber);
 		extraData.setNonce_str(String.valueOf(System.currentTimeMillis()));
 		Map<String, String> map = MapUtil.objectToMap(extraData);
 		extraData.setSign(SignatureUtil.generateSign(map, key));
@@ -116,9 +142,9 @@ public class PayVeticleAPI extends BaseAPI {
 	 * @param oldAttach
 	 * @param plateNumber
 	 * @param appid
-	 * @return
+	 * @return 下单时附加字段attach的字符串
 	 */
-	public static String openServiceWhenPaying(String oldAttach, String plateNumber, String appid) {
+	public static String getAttachToOpenServiceOnPaying(String oldAttach, String plateNumber, String appid) {
 		return String.format("#*#{\"pn\":\"%s\",\"aid\":\"%s\"}#*#%s", plateNumber, appid, oldAttach);
 	}
 }
