+++ src/main/java/weixin/popular/bean/vehicle/PartnerpayPayApplyResult.java
@@ -1,0 +1,27 @@
+package weixin.popular.bean.vehicle;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlRootElement;
+
+/**
+ * 申请扣款结果
+ * 
+ * @author around
+ *
+ */
+@XmlRootElement(name = "xml")
+@XmlAccessorType(XmlAccessType.FIELD)
+public class PartnerpayPayApplyResult extends VehicleBaseResult {
+
+	private String device_info;
+
+	public String getDevice_info() {
+		return device_info;
+	}
+
+	public void setDevice_info(String device_info) {
+		this.device_info = device_info;
+	}
+
+}
