+++ src/main/java/weixin/popular/bean/vehicle/PartnerpayQueryState.java
@@ -12,7 +12,7 @@ import javax.xml.bind.annotation.XmlRootElement;
  */
 @XmlRootElement(name = "xml")
 @XmlAccessorType(XmlAccessType.FIELD)
-public class QueryState extends VehicleBaseReq {
+public class PartnerpayQueryState extends VehicleBaseReq {
 
 	private String openid;
