+++ src/main/java/weixin/popular/bean/vehicle/PartnerpayQueryStateResult.java
@@ -20,17 +20,7 @@ import weixin.popular.bean.DynamicField;
  */
 @XmlRootElement(name = "xml")
 @XmlAccessorType(XmlAccessType.FIELD)
-public class QueryStateResult extends VehicleBaseResult implements DynamicField {
-
-	private String appid;
-	private String sub_appid;
-	private String mch_id;
-	private String sub_mch_id;
-	private String nonce_str;
-	private String sign;
-	private String result_code;
-	private String err_code;
-	private String err_code_des;
+public class PartnerpayQueryStateResult extends VehicleBaseResult implements DynamicField {
 
 	private String user_state;
 	private String openid;
@@ -45,79 +35,10 @@ public class QueryStateResult extends VehicleBaseResult implements DynamicField 
 	// 车牌号列表。仅包括省份+车牌，不包括特殊字符。多个车牌时，以分好间隔。
 	private String plate_number_list;// v2.0前
 	private List<PlateNumberInfo> plate_number_info;// v2.0以后返回这个
-
-	public String getAppid() {
-		return appid;
-	}
-
-	public void setAppid(String appid) {
-		this.appid = appid;
-	}
-
-	public String getSub_appid() {
-		return sub_appid;
-	}
-
-	public void setSub_appid(String sub_appid) {
-		this.sub_appid = sub_appid;
-	}
-
-	public String getMch_id() {
-		return mch_id;
-	}
-
-	public void setMch_id(String mch_id) {
-		this.mch_id = mch_id;
-	}
-
-	public String getSub_mch_id() {
-		return sub_mch_id;
-	}
-
-	public void setSub_mch_id(String sub_mch_id) {
-		this.sub_mch_id = sub_mch_id;
-	}
-
-	public String getNonce_str() {
-		return nonce_str;
-	}
-
-	public void setNonce_str(String nonce_str) {
-		this.nonce_str = nonce_str;
-	}
-
-	public String getSign() {
-		return sign;
-	}
-
-	public void setSign(String sign) {
-		this.sign = sign;
-	}
-
-	public String getResult_code() {
-		return result_code;
-	}
-
-	public void setResult_code(String result_code) {
-		this.result_code = result_code;
-	}
-
-	public String getErr_code() {
-		return err_code;
-	}
-
-	public void setErr_code(String err_code) {
-		this.err_code = err_code;
-	}
-
-	public String getErr_code_des() {
-		return err_code_des;
-	}
-
-	public void setErr_code_des(String err_code_des) {
-		this.err_code_des = err_code_des;
-	}
-
+	
+	// 3.0新增
+	private String deduct_mode;
+	
 	public String getUser_state() {
 		return user_state;
 	}
@@ -174,5 +95,13 @@ public class QueryStateResult extends VehicleBaseResult implements DynamicField 
 			JSONObject info = JSON.parseObject(json);
 			this.plate_number_info = JSON.parseArray(info.getString("plate_number_info"), PlateNumberInfo.class);
 		}
+	}
+
+	public String getDeduct_mode() {
+		return deduct_mode;
+	}
+
+	public void setDeduct_mode(String deduct_mode) {
+		this.deduct_mode = deduct_mode;
 	}
 }
