+++ src/main/java/weixin/popular/bean/vehicle/PlateNumberInfo.java
@@ -9,6 +9,8 @@ package weixin.popular.bean.vehicle;
 public class PlateNumberInfo {
 	private String plate_number;
 	private String channel_type;
+	// 3.0新增
+	private String common_use_flag;
 
 	public PlateNumberInfo() {
 	}
@@ -32,5 +34,13 @@ public class PlateNumberInfo {
 
 	public void setChannel_type(String channel_type) {
 		this.channel_type = channel_type;
+	}
+
+	public String getCommon_use_flag() {
+		return common_use_flag;
+	}
+
+	public void setCommon_use_flag(String common_use_flag) {
+		this.common_use_flag = common_use_flag;
 	}
 }
